Refactor for NM
